* commit '483a3da9eac6046d599e5ca0f6602ff86348e4dd':
  Transacties krijgen ook een bedrag mee, wel zo handig. Saldo's zijn nu floats. De ing werkt nog steeds niet

+++ src/main/java/API/DTO/BalanceAmount.java
@@ -2,7 +2,7 @@ package API.DTO;
 
 public class BalanceAmount {
     private String currency;
-    private int amount;
+    private float amount;
 
     public BalanceAmount() {
     }
@@ -20,11 +20,11 @@ public class BalanceAmount {
         this.currency = currency;
     }
 
-    public int getAmount() {
+    public float getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
 }
